requirements.txt
streamlit==1.32.0
openai==1.13.3
pinecone-client==3.1.0
langchain==0.1.11
langchain-pinecone==0.0.3
langchain-openai==0.0.8